6: relocate the handler for `POST /api/sign_in` requests to `src/auth_subsystem/views.py`
