+++ src/auth_subsystem/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path(
+        "sign_in",
+        views.sign_in,
+        name="sign-in",
+    ),
+]
