+++ src/mini_jira_2/urls.py
@@ -20,4 +20,5 @@ from django.urls import path, include
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/", include("src.tasks.urls")),
+    path("api/", include("src.auth_subsystem.urls")),
 ]
