+++ src/tasks/urls.py
@@ -12,9 +12,4 @@ urlpatterns = [
         views.process_task,
         name="process-task",
     ),
-    path(
-        "sign_in",
-        views.sign_in,
-        name="sign-in",
-    ),
 ]
